requirements.txt
vk==3.0
python-dotenv==0.21.0
